requirements.txt
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-cors-headers==2.5.0
django-debug-toolbar==1.11
django-friendship==1.8.0
django-rest-auth==0.9.3
djangorestframework==3.9.2
geographiclib==1.49
geoip2==2.9.0
geopy==1.19.0
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
idna==2.8
maxminddb==1.4.1
oauth2client==4.1.3
oauthlib==3.0.1
Pillow==5.4.1
psycopg2-binary==2.7.7
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycrypto==2.6.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==3.3.1
